Improved profiling support in psort
